Add `DESTDIR/share` to paths to look into for Nautilus plugin folder
In some package formats (like nixpkg), the paths that we're looking into
for the `nautilus-python` directory may not be enough, so add another
one based on the DESTDIR (e.g., `DESTDIR/share`) that's set during the
meson build.

Here is a snippet where the package maintainers are manually symlinking
the plugin script to the required directory since the app is not able to
figure it out by itself.
https://github.com/NixOS/nixpkgs/pull/416076/files#diff-2b073efb0973697970f4ba24dec07b65f7aea950aa3f48ba4f2d4a92827ac998R74-R76